--- src/common/message-creators/istore-action-message-creator.d.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-export interface StoreActionMessageCreator {
-    getAllStates: () => void;
-}
