+++ src/tests/unit/tests/common/message-creators/store-action-message-creator.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { IMock, It, Mock, MockBehavior } from 'typemoq';
 
-import { StoreActionMessageCreatorImpl } from '../../../../../common/message-creators/store-action-message-creator';
+import { StoreActionMessageCreatorImpl } from '../../../../../common/message-creators/store-action-message-creator-impl';
 
 describe('StateActionMessageCreator', () => {
     const tabId: number = -1;
